Module rustopals::block

Block ciphers implementations and related utilities.

Re-exports

pub use aes128::AES128;

pub use cbc::CBC;

pub use ecb::ECB;

pub use pkcs7::PKCS7Error;

Modules

aes128
AES-128 block-cipher.

cbc
CBC block mode, where each block of plaintext is XORed with the previous ciphertext block before being encrypted.

ecb
ECB block mode, where each block is encrypted separately.

pkcs7
PKCS7 padding scheme. Pads/unpads plaintext/ciphertext to a multiple of the block size.

Enums

Mode
Block-cipher mode of operation.

Traits

BlockCipher
Trait for block ciphers.

BlockMode
Trait for block-cipher modes of operation

Functions

count_repeated
Count repeated block_size bocks in data.