Module rustopals::block::pkcs7

PKCS7 padding scheme. Pads/unpads plaintext/ciphertext to a multiple of the block size.

The padding byte value (repeated for each byte of padding) is the number of bytes that are added, (i.e. N bytes of value N are added). This implies padding values are <= block_size.

If no padding is needed, a full block (i.e. block_size bytes) of value block_size is added. This is required to disambiguate padding from an actual ending byte. This implies 0x00 padding is invalid.

Enums

PKCS7Error
Possible PKCS7 padding errors.

Functions

pad
Immutably pads payload to a multiple of block_length. Returns a new buffer.

pad_vec
Mutably pads payload to a multiple of block_length. Modifies the payload buffer.

unpad
Immutably unpads payload from a multiple of block_length. Returns a slice of the original buffer.

unpad_vec
Mutably unpads payload from a multiple of block_length. Modifies the payload buffer.