Trait rustopals::digest::Digest

pub trait Digest: Default {
    type Output: AsRef<[u8]>;

    const OUTPUT_LENGTH: usize;
    const BLOCK_LENGTH: usize;
    const ASN1_PREFIX: &'static [u8];

    fn update(&mut self, message: &[u8]);

    fn finalize(self) -> Self::Output;

    fn chain(self, message: &[u8]) -> Self
    where
        Self: Sized,
    { ... }

    #[must_use]
    fn digest(message: &[u8]) -> Self::Output
    where
        Self: Default,
    { ... }
}

Trait for message digest implementations.

Associated Types

type Output: AsRef<[u8]>

Loading content...
Associated Constants

const OUTPUT_LENGTH: usize

const BLOCK_LENGTH: usize

const ASN1_PREFIX: &'static [u8]

Loading content...
Required methods

fn update(&mut self, message: &[u8])

Update the digest with message bytes

fn finalize(self) -> Self::Output

Finalize the digest and get its value

Loading content...
Provided methods

fn chain(self, message: &[u8]) -> Self where
Self: Sized,

Convenience method to update the digest with message bytes in a chainable fashion

#[must_use]
fn digest(message: &[u8]) -> Self::Output where
Self: Default,

Convenience method to create, update and finalize the digest with message bytes

Loading content...