Trait rustopals::digest::ExtensibleDigest

pub trait ExtensibleDigest: Digest + Sized {
    fn extend_digest(
        digest_output: Self::Output, 
        guessed_payload_length: usize
    ) -> (Self, Vec<u8>);
}

Trait for digests that can be subject to length-extension attacks.

Useful for cracking bad MAC constructs like naive key-prefixed MACs.

Required methods

fn extend_digest(
digest_output: Self::Output,
guessed_payload_length: usize
) -> (Self, Vec<u8>)

Perform a length-extension attack on an existing digest result.

Given:

digest_output: A known digest that we want to extend.

guessed_payload_length: The guessed length of what has been digested so far (not including the padding).

It returns:

The reverse-engineered Digest that we can extend.

The synthetic padding as a Vec<u8> that we have to append to our plaintext before extending it.

Loading content...